requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
frozenlist==1.3.0
geoip2==4.6.0
idna==3.3
iniconfig==1.1.1
maxminddb==2.2.0
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.10
wikipedia==1.4.0
yarl==1.7.2
django-dotenv==1.4.2
